@@ -11,7 +11,7 @@
 #  License: MIT (see LICENSE file)
 
 from datetime import date
-from typing import Optional, Tuple
+from typing import Optional
 
 from holidays.calendars import _HinduLunisolar
 
@@ -26,7 +26,7 @@ class HinduCalendarHolidays:
         self._hindu_calendar_show_estimated = show_estimated
 
     def _add_hindu_calendar_holiday(
-        self, name: str, dt_estimated: Tuple[date, bool]
+        self, name: str, dt_estimated: tuple[date, bool]
     ) -> Optional[date]:
         """
         Add Hindu calendar holiday.
