@@ -10,8 +10,8 @@
 #  Website: https://github.com/vacanza/holidays
 #  License: MIT (see LICENSE file)
 
+from collections.abc import Iterable
 from datetime import date
-from typing import Iterable, Set, Tuple
 
 from holidays.calendars import _IslamicLunar
 from holidays.calendars.gregorian import _timedelta
@@ -28,7 +28,7 @@ class IslamicHolidays:
     def __init__(self, cls=None) -> None:
         self._islamic_calendar = cls() if cls else _IslamicLunar()
 
-    def _add_ali_al_rida_death_day(self, name) -> Set[date]:
+    def _add_ali_al_rida_death_day(self, name) -> set[date]:
         """
         Add death of Ali al-Rida day (last (29th or 30th) day of 2nd month).
 
@@ -38,7 +38,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.ali_al_rida_death_dates(self._year)
         )
 
-    def _add_ali_birthday_day(self, name) -> Set[date]:
+    def _add_ali_birthday_day(self, name) -> set[date]:
         """
         Add birthday of Ali ibn Abu Talib day (13th day of 7th month).
 
@@ -48,7 +48,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.ali_birthday_dates(self._year)
         )
 
-    def _add_ali_death_day(self, name) -> Set[date]:
+    def _add_ali_death_day(self, name) -> set[date]:
         """
         Add death of Ali ibn Abu Talib day (21st day of 9th month).
 
@@ -58,7 +58,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.ali_death_dates(self._year)
         )
 
-    def _add_arbaeen_day(self, name) -> Set[date]:
+    def _add_arbaeen_day(self, name) -> set[date]:
         """
         Add Arbaeen day (20th day of 2nd month).
 
@@ -68,7 +68,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.arbaeen_dates(self._year)
         )
 
-    def _add_arafah_day(self, name) -> Set[date]:
+    def _add_arafah_day(self, name) -> set[date]:
         """
         Add Day of Arafah (9th day of 12th month).
 
@@ -81,7 +81,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.eid_al_adha_dates(self._year), days_delta=-1
         )
 
-    def _add_ashura_day(self, name) -> Set[date]:
+    def _add_ashura_day(self, name) -> set[date]:
         """
         Add Ashura Day (10th day of 1st month).
 
@@ -93,7 +93,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.ashura_dates(self._year)
         )
 
-    def _add_ashura_eve(self, name) -> Set[date]:
+    def _add_ashura_eve(self, name) -> set[date]:
         """
         Add Ashura Eve (Day before the 10th day of 1st month).
 
@@ -105,7 +105,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.ashura_dates(self._year), days_delta=-1
         )
 
-    def _add_eid_al_adha_day(self, name) -> Set[date]:
+    def _add_eid_al_adha_day(self, name) -> set[date]:
         """
         Add Eid al-Adha Day (10th day of the 12th month of Islamic calendar).
 
@@ -118,7 +118,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.eid_al_adha_dates(self._year)
         )
 
-    def _add_eid_al_adha_day_two(self, name) -> Set[date]:
+    def _add_eid_al_adha_day_two(self, name) -> set[date]:
         """
         Add Eid al-Adha Day Two.
 
@@ -128,7 +128,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.eid_al_adha_dates(self._year), days_delta=+1
         )
 
-    def _add_eid_al_adha_day_three(self, name) -> Set[date]:
+    def _add_eid_al_adha_day_three(self, name) -> set[date]:
         """
         Add Eid al-Adha Day Three.
 
@@ -138,7 +138,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.eid_al_adha_dates(self._year), days_delta=+2
         )
 
-    def _add_eid_al_adha_day_four(self, name) -> Set[date]:
+    def _add_eid_al_adha_day_four(self, name) -> set[date]:
         """
         Add Eid al-Adha Day Four.
 
@@ -148,7 +148,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.eid_al_adha_dates(self._year), days_delta=+3
         )
 
-    def _add_eid_al_fitr_day(self, name) -> Set[date]:
+    def _add_eid_al_fitr_day(self, name) -> set[date]:
         """
         Add Eid al-Fitr Day (1st day of 10th month of Islamic calendar).
 
@@ -161,7 +161,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.eid_al_fitr_dates(self._year)
         )
 
-    def _add_eid_al_fitr_day_two(self, name) -> Set[date]:
+    def _add_eid_al_fitr_day_two(self, name) -> set[date]:
         """
         Add Eid al-Fitr Day Two.
 
@@ -171,7 +171,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.eid_al_fitr_dates(self._year), days_delta=+1
         )
 
-    def _add_eid_al_fitr_day_three(self, name) -> Set[date]:
+    def _add_eid_al_fitr_day_three(self, name) -> set[date]:
         """
         Add Eid al-Fitr Day Three.
 
@@ -181,7 +181,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.eid_al_fitr_dates(self._year), days_delta=+2
         )
 
-    def _add_eid_al_fitr_day_four(self, name) -> Set[date]:
+    def _add_eid_al_fitr_day_four(self, name) -> set[date]:
         """
         Add Eid al-Fitr Day Four.
 
@@ -191,7 +191,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.eid_al_fitr_dates(self._year), days_delta=+3
         )
 
-    def _add_eid_al_fitr_eve(self, name) -> Set[date]:
+    def _add_eid_al_fitr_eve(self, name) -> set[date]:
         """
         Add Eid al-Fitr Eve (last day of 9th month of Islamic calendar).
 
@@ -201,7 +201,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.eid_al_fitr_dates(self._year), days_delta=-1
         )
 
-    def _add_eid_al_ghadir_day(self, name) -> Set[date]:
+    def _add_eid_al_ghadir_day(self, name) -> set[date]:
         """
         Add Eid al-Ghadir Day (18th day of 12th month).
 
@@ -211,7 +211,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.eid_al_ghadir_dates(self._year)
         )
 
-    def _add_fatima_death_day(self, name) -> Set[date]:
+    def _add_fatima_death_day(self, name) -> set[date]:
         """
         Add death of Fatima day (3rd day of 6th month).
 
@@ -221,7 +221,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.fatima_death_dates(self._year)
         )
 
-    def _add_hari_hol_johor(self, name) -> Set[date]:
+    def _add_hari_hol_johor(self, name) -> set[date]:
         """
         Hari Hol Johor.
 
@@ -231,7 +231,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.hari_hol_johor_dates(self._year)
         )
 
-    def _add_hasan_al_askari_death_day(self, name) -> Set[date]:
+    def _add_hasan_al_askari_death_day(self, name) -> set[date]:
         """
         Add death of Hasan_al-Askari day (8th day of 3rd month).
 
@@ -241,7 +241,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.hasan_al_askari_death_dates(self._year)
         )
 
-    def _add_imam_mahdi_birthday_day(self, name) -> Set[date]:
+    def _add_imam_mahdi_birthday_day(self, name) -> set[date]:
         """
         Add birthday of Muhammad al-Mahdi day (15th day of 8th month).
 
@@ -252,8 +252,8 @@ class IslamicHolidays:
         )
 
     def _add_islamic_calendar_holiday(
-        self, name: str, dates: Iterable[Tuple[date, bool]], days_delta: int = 0
-    ) -> Set[date]:
+        self, name: str, dates: Iterable[tuple[date, bool]], days_delta: int = 0
+    ) -> set[date]:
         """
         Add lunar calendar holiday.
 
@@ -274,7 +274,7 @@ class IslamicHolidays:
 
         return added_dates
 
-    def _add_islamic_new_year_day(self, name) -> Set[date]:
+    def _add_islamic_new_year_day(self, name) -> set[date]:
         """
         Add Islamic New Year Day (last day of Dhu al-Hijjah).
 
@@ -299,7 +299,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.isra_and_miraj_dates(self._year)
         )
 
-    def _add_maldives_embraced_islam_day(self, name) -> Set[date]:
+    def _add_maldives_embraced_islam_day(self, name) -> set[date]:
         """
         Add Maldives Embraced Islam Day (1st day of 4th month).
 
@@ -309,7 +309,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.maldives_embraced_islam_day_dates(self._year)
         )
 
-    def _add_mawlid_day(self, name) -> Set[date]:
+    def _add_mawlid_day(self, name) -> set[date]:
         """
         Add Mawlid Day (12th day of 3rd month).
 
@@ -321,7 +321,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.mawlid_dates(self._year)
         )
 
-    def _add_mawlid_day_two(self, name) -> Set[date]:
+    def _add_mawlid_day_two(self, name) -> set[date]:
         """
         Add Mawlid Day Two.
 
@@ -333,7 +333,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.mawlid_dates(self._year), days_delta=+1
         )
 
-    def _add_nuzul_al_quran_day(self, name) -> Set[date]:
+    def _add_nuzul_al_quran_day(self, name) -> set[date]:
         """
         Add Nuzul Al Quran (17th day of 9th month).
 
@@ -346,7 +346,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.nuzul_al_quran_dates(self._year)
         )
 
-    def _add_prophet_death_day(self, name) -> Set[date]:
+    def _add_prophet_death_day(self, name) -> set[date]:
         """
         Add death of Prophet Muhammad and Hasan ibn Ali day (28th day of 2nd month).
 
@@ -356,7 +356,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.prophet_death_dates(self._year)
         )
 
-    def _add_quamee_dhuvas_day(self, name) -> Set[date]:
+    def _add_quamee_dhuvas_day(self, name) -> set[date]:
         """
         Add Quamee Dhuvas (1st day of 3rd month).
 
@@ -366,7 +366,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.quamee_dhuvas_dates(self._year)
         )
 
-    def _add_ramadan_beginning_day(self, name) -> Set[date]:
+    def _add_ramadan_beginning_day(self, name) -> set[date]:
         """
         Add First Day of Ramadan (1st day of 9th month).
 
@@ -378,7 +378,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.ramadan_beginning_dates(self._year)
         )
 
-    def _add_sadiq_birthday_day(self, name) -> Set[date]:
+    def _add_sadiq_birthday_day(self, name) -> set[date]:
         """
         Add birthday of Prophet Muhammad and Ja'far al-Sadiq day (17th day of 3rd month).
 
@@ -388,7 +388,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.sadiq_birthday_dates(self._year)
         )
 
-    def _add_sadiq_death_day(self, name) -> Set[date]:
+    def _add_sadiq_death_day(self, name) -> set[date]:
         """
         Add death of Ja'far al-Sadiq day (25th day of 10th month).
 
@@ -398,7 +398,7 @@ class IslamicHolidays:
             name, self._islamic_calendar.sadiq_death_dates(self._year)
         )
 
-    def _add_tasua_day(self, name) -> Set[date]:
+    def _add_tasua_day(self, name) -> set[date]:
         """
         Add Tasua day (9th day of 1st month).
 
